# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookVerse API Documentation",
        "description": "A comprehensive API for managing books, orders, payments, and AI-powered features",
        "contact": {
            "name": "BookVerse Support",
            "email": "support@bookverse.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "BookVerse API Server"
        }
    ],
    "paths": {
...
        "/api/books/{id}/recommendations": {
            "get": {
                "tags": [
                    "AI Features"
                ],
                "summary": "Get AI recommendations for a book",
                "description": "Get AI-powered book recommendations based on the specified book",
                "operationId": "769c9292735ca618e569b88c2480e67b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendations response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Recommendations retrieved successfully."
                                                },
                                                "recommendations": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "To Kill a Mockingbird"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "Harper Lee"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "A powerful story of racial injustice..."
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 15.99
                                                            },
                                                            "similarity_score": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 0.85
                                                            },
                                                            "reason": {
                                                                "type": "string",
                                                                "example": "Similar themes and writing style"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Generating recommendations. Please try again in a few moments."
                                                },
                                                "recommendations": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}